requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.13.0
djangorestframework==3.14.0
django-crontab==0.7.1
pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
requests==2.28.2
xlrd==2.0.1
xlwt==1.3.0
mysqlclient==2.1.1
